Generalize Cluster Resources
