+++ cmd/kops/delete_cluster.go
@@ -18,7 +18,6 @@ package main
 
 import (
 	"fmt"
-
 	"io"
 
 	"github.com/golang/glog"
@@ -143,7 +142,7 @@ func RunDeleteCluster(f *util.Factory, out io.Writer, options *DeleteClusterOpti
 		}
 
 		// Todo lets make this smart enough to detect the cloud and switch on the ClusterResources interface
-		d := &resources.AwsCluster{}
+		d := &resources.ClusterResources{}
 		d.ClusterName = clusterName
 		d.Cloud = cloud
