+++ pkg/resources/aws_test.go
